Clean up context-specific options
The big conditional near the top of the file was full of stuff that was
commented out because it didn't work or was set on a different OS or
environment and didn't work in the current one.

Using more fine grained conditionals I've been able to re-enable all
options so that they only run in the contexts they're supposed to, like
choosing the right font for Windows™ or the right listchars for TTY
terminals. As a nice side-effect, PowerLine is used in all terminals.