Fixed problem of doc tags dirtying sub-modules.
Some sub-modules' gitignore files don't tell git to ignore doc/tags.
Previously I tried solving this by adding this to the sub-module's
gitignore file and commiting that change both there and in my main
repository but this causes problems because I can't (and didn't) push to
those repositories, so when trying to sync those sub-modules on another
machine, git fails because my repository references non-existing
commits.

I've fixed this by undoing those commits (the offending repos' HEADs are
now at the tips of their master branches again) and adding an ignore
line for each sub-module in the gitmodules file. In future I should
probably script adding sub-modules because it's going to be annoying if
I have to add that line each time I add a sub-module.