Add CD plugin to ease navigation
The CD plugin adds a CD command which takes you to the directory of the
current file. Unfortunately in the latest version this functionality has
been changed to remove the command and make the plugin change directory
automatically, which is not what I want. Therefore, I have checked out
version 1.1, the latest version before this change in functionality was
introduced.