Sets buffers to delete when editing TeX files
In vimrc buffers are set to hide instead of delete when closing with the
command `set hidden`. This causes problems with latex-suite's error log
and quick-fix list, which seem to assume that the previous list/log
buffer has been deleted before it shows the next one, which causes
errors. Turning off "hidden" for .tex files solves this.