Change modeline precedes and extends chars
The precedes and extends characters were ← and → respectively, but → is
longer than the width of a character (at least in the Inconsolata font)
and so the arrow-head will appear off the edge of the screen where it
can't be seen. I've changed the arrows to guillemets « and », which fit
in with the symbol for tabs and don't have the problems the arrows did.