Change submodule URLs to use HTTPS transport instead of SSH
GitHub expects the user to authenticate with an SSH keypair when using
the SSH protocol, but with the HTTPS protocol GitHub behaves 'smartly'
and lets the user access it read-only if they don't authenticate or
don't have write access. This is nice because people without a github
account can now clone the repo without having to fiddle with the URLs.