Add an exception for git commits to the line recalling autocmd
Commit 26cffeabeb276d5ea03e0bdd46443c0cf5ad3e8e added an autocmd to
place the cursor at the start of the file for git commits, in order to
override an earlier autocmd which recalls the last cursor position.
Using a tip from the net, I update removed the second autocmd and
instead updated the first one to make an exception for git commit
messages.