Add proper highlighting and IDE functions for Arduino
Instead of just using the C++ syntax files, I added a syntax script by
Johannes Hoff which takes the C++ syntax and adds Arduino specific
keywords. I added this to tclem's vim-arduino plugin which lets you
compile and deploy Arduino sketches from vim and added this combination
to my pathogen bundles.

This should make Arduino development a bit more fun as the IDE was a bit
annoying to work with.