Make simplified vim colours/chars work in screen and drop-down term
  * in GNU Screen &term is screen
  * in GNOME drop-down terminal &term is xterm

I've added these to the OR list of terminals for which Vim should use
simplified characters and colours. Actually this should also be good if
I run it in an actual xterm as well because I don't think xterm is
likely to be able to handle these.

I also reformatted the list with one OR per line because the line was
getting too long.