Set tex_flavor to latex in vimrc
Previously I moved all LaTeX settings to an ftplugin but this one needs
to be in the vimrc because it forces the filetype of .tex files to be
tex instead of plaintex. Obviously plaintex files wouldn't load the tex
ftplugin and all the latex-suite stuff doesn't load for them anyway
which sucks. With this in the vimrc everything works as expected.