Fix the latex cleanup AutoCommand
I had an AutoCommand to clean up rubbish like auxiliary files and logs
files strewn by pdflatex when compiling PDFs which I wrote for bash on
Windows. It runs when closing the tex file, because those are actually
needed for compiling the PDFs and it's better if they persist while
you're editing them.

When I wrote it, it seemed to work but only in the current directory. On
Linux I noticed that interestingly it was trying to delete these rubbish
files in the home directory instead of in the directory I was working
in, which is a) useless and b) dangerous. I read up the vim docs for %
and rewrote the clean up command with grep and xargs, so it also now no
longer whines if the files don't exist when it tries to delete them.

The regex catches the most common rubbish files but it might have to be
extended later if I work with weirder tex files (I'm thinking .lol for
List of Lists or something).