Adjust GVim window size
Since I tend to use a line width around 80 in most files, it doesn't
make sense making Vim wider than that. By default it's now narrower,
more portrait-like, instead of filling up so much of my screen with
white space.

In future it might be nice to add an autocommand or something that makes
GVim wider when I open a diff or a vertical split.