Improved colouring for black terminals
Xterm isn't bad, you just need to tell Solarized to leave the background
transparent to let it look OK.  Specifically, the extension-provided
GNOME drop-down terminal's background is black, it reports itself as
xterm and I use it a lot.  Everything looks fine with it this way,
although I'm curious how it'll be with xterm's default white background.