Massive refactor of setting font face/size based on OS
I had enough of commenting out and rewriting the guifont setting on
every new machine.

Now everywhere where fonts were set, it just sets what OS you're on and
later on the script looks it up in a hash of OS-to-fonts together with a
default font size.  The advantage is you don't have to specify the full
font name with size and there's a helper function which sets the size,
keeping the font the same.  Because the helper also tracks the size
locally, it was trivial to add commands and mappings for font size
increment and decrement.  This is a long itch I've just scratched.