Use yadm for fugitive diffs when you're in the home directory
YADM tracks your dotfiles in home in git, it's really annoying to
compare patches without fugitive.  This setting tells fugitive to use
yadm instead of git, which means the right git status, diff, etc will be
shown.  In truth fugitive doesn't really seem to use this value much, so
I'm working on a patch for that.