Add Gist plugin for easy access to GitHub gists
This depends on the webapi plugin from the same author, hopefully it'll
be useful for something else later too and not just this one plugin.

As far as I can tell you can't get it to support multiple GitHub's, like
Fugitive does, so I've set it up to Schibsted's GitHub, which I use more
often than github.com.