Simplify wiki paths
All those ifs lead to cryptic error messages that don't really tell you
why the wiki isn't working.  I've moved the wiki source to an easier to
reach place and expect it to always exist on my personal machine.  If it
doesn't exist, vim will complain when you try to use the wiki and it
will be clear what needs to be done.