Remove hard coded GUI window size
Tiling window managers don't respect this by design, and make the window
whatever size it needs to be, this just causes weird stuff to happen,
like the drawable part of the window being a tiny slice of the full
window with the rest greyed out, or changing the font size causing the
window size to halve... wtf... best not to set this.

If I go back to stacking window managers some day this might be useful
again, but preferably in a more dynamic way.