Automatically clear screen when saving in :Goyo
Without this every time I save the file it leaves the "bytes written"
message at the bottom.  I often use Goyo to present the screen in video
calls because it looks cleaner and less distracting than full vim but I
had to always press Ctrl-L as a reflex after each save.  This does that
for me.