Add Join plugin and quick visual mapping to use it
The use case is:  I want something just like J but with comma separation
instead of just spaces.  It feels a bit OTT to pull in a whole plugin
just for this but I'll see if I use its features and if not then I'll
probably write something that does only this one thing that I want.