Turn off clearing status messages after exiting Goyo
There was a cool trick with CursorHold * :echo that would echo an empty
message to clear the text at the bottom of the screen e.g. whenever you
write a file.  This is slick in distraction-free writing mode with Goyo,
which I often also use for presenting, because then this ugly,
distracting text isn't shown for more than a fraction of a second.

The problem is once you come out of Goyo this behaviour still sticks and
then I can't see stuff that was echoed out intentionally.  Anyway, the
solution is simple, put this autocommand inside an augrounp (I've named
it distractionfree) and then that can be used to delete all the
autocommands in it when you leave Goyo (so far only this one).