Use explicit fullscreen on/off commands when entering/leaving Goyo
It's cool that it'll toggle fullscreen mode for distraction free writing
but it'll get mixed up if you manually take vim out of fullscreen mode
and then leave Goyo, it'll toggle you back into fullscreen.  The inverse
is just as silly.

So, I checked the docs for wmctrl and apart from toggle you can also add
and remove properties, so I've used these to make explicit FullScreenOn
and FullScreenOff commands and use these within the goyo_enter and
goyo_leave callbacks instead of FullScreenToggle.  I also changed the
order between changing to fullscreen and changing the font size, to
avoid it making this weird tiny big-sized window with lots of space on
the sides.