Add hack for removing titlebar in Goyo
Sometimes I don't use Goyo's distraction-free writing mode just to keep
myself focussed when writing things, but because it looks really good
when presenting or taking notes live while sharing the screen.  I take
my notes in Vimwiki anyway, but hiding the statusline and line numbers
etc. makes it looks nice and minimal, especially together with the nice
layout Goyo provides.  Anyway, so far I had Goyo also go fullscreen but
this isn't great if I'm presenting from a laptop because then I can't
see or interact with the ongoing call I'm presenting in, so I've
commented out that fullscreen part.  However I thought it looked ugly
having the Window titlebar shown in the screencast a) wasting space and
b) saying "GVIM - diary bla bla bla".  Not professional.

I started looking for a way to hide the titlebar for Gvim or for apps
generally in GNOME and didn't find much.  There are a bunch of GNOME
extensions emulating Unity look-and-feel that move Window titles to the
top bar but they're no longer maintained and don't match this GNOME
version, wmctrl (that I use for fullscreening) doesn't seem to support a
titlebar-related prop, most of the posts are explaining about how to
toggle the _toolbar_ but that is trivial.  Eventually I found these:

https://unix.stackexchange.com/q/156300/
https://gist.github.com/zenozeng/8897839

Putting two and two together I figured using xprop, the 3rd parameter
should be 0x0 for hiding the titlebar and 0x1 for showing it again.
Fantastic, so I wrote a function that handles that and includes some
internal state for toggling and use that now in Goyo.  Works exactly as
I had hoped.