Resize to a small square window when leaving Goyo
The font size changes and fullscreen toggling seem to mess with the
window size and keeps getting bigger each time.  To keep the window
reasonably sized after leaving, I've hardcoded a good-looking window
size, I'm not sure it's 100% a size I'll like but it seems alright.
Obviously the disadvantage is that whatever the window was sized to
before will be totally lost, but that was kind of already the case.