Fix spelling underwave/underline in Terminal
I spent so long searching why this happens, I don't really get it,
something is messed up with terminal capability reporting and how Vim
handle it and that's why it hasn't been catching typos in my commit
messages for the last 2 years.  Hopefully this snippet I copied off a
GitHub comment solves it not just on this 1 computer cos right now not
only is the spelling wiggle bag, but some extra things have appeared
that are new, like a small notch pointing at where the error is in my
code within the reported line.  Amazing.