Add SwitchDayNightTheme command to switch to dark/light
I wrapped the existing code that checked GNOME Settings to see whether
it was using dark mode or light mode and set the vim background
accordingly with a function.  The function has a command associated with
it: :SwitchDayNightTheme.  I also wanted to add an autocmd that would do
this switch automatically for some event like when the window gets focus
but that didn't seem to work.