Reorder Day/Night theme switcher to set background before theme
For some reason in Gvim I experienced an issue with the current order,
so it'd set the theme correctly, and then afterwards when setting the
background to dark/light it would revert to the 'default' theme which
looks lame.  By re-ordering it so that the background setting comes
first and the colortheme last this doesn't happen and I get the desired
catppuccin theme.