Update ultisnips and snippets to the latest version
Critically, the update to the honza/vim-snippets repo fixes a character
encoding issue that was constantly spamming while using snippets.
Luckily I found the issue on GitHub, it's pretty recent[1], otherwise I
was gearing up to switch to neosnippets, nothing wrong with it, it might
even be better, cool, modern async stuff, but it's not what's in my
muscle memory.

[1]: https://github.com/honza/vim-snippets/issues/1500